LayoutTests/http/tests/security/xss-DENIED-script-inject-into-inactive-window2-pson-expected.txt

CONSOLE MESSAGE: Victim loaded.
This page doesn't do anything special.
